+++ LoL/Controllers/TeamsController.cs
@@ -150,6 +150,9 @@ namespace LoL.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var team = await _context.Teams.FindAsync(id);
+            if (team is null)
+                return NotFound();
+
             _context.Teams.Remove(team);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
